folded scenemanager logic into main menu and gamemanger scripts
