+++ Assets/Scripts/MainMenu.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 206a8bdfed929f945bf0095cc8c1ea71
+guid: 30ab7781aeeca08438b2b14fb879955a
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
